check existence

+++ update_local_conf.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
 for file in gitconfig vimrc bash_aliases bashrc; do
-  cp -f $HOME/.$file{,.backup-`date +%Y%m%d-%H%M`}
+  if test -f $HOME/.$file; then
+    cp -f $HOME/.$file{,.backup-`date +%Y%m%d-%H%M`}
+  fi
   cp -f {,$HOME/}.$file
 done
